Fix default values overwriting data passed
In the "Set default objects with Object.assign" section your example would overwrite the data passed as a parameter with the default values. Changing the parameter order fixes the issue.

Also, the original version was mutating the passed object, but this is non-issue now.

Thanks to https://www.reddit.com/user/notNullOrVoid for spotting the mixed order.