Replace pass-by-reference with mutability, clean up
This edit only affects the Avoid Side Effects (part 2) section.
It replaces the discussion of pass by reference with an explanation in terms of mutability. 
By avoiding the confusing terminology and possible links to the implementation details of 
different JS engines the explanation can be kept at an abstraction level that's great for 
beginners and experts alike. I also cleaned up some grammar, typos, etc.